Small fixes.
