+++ app/src/main/res/layout/fragment_maps.xml
@@ -28,7 +28,7 @@
             app:layout_constraintTop_toTopOf="parent" />
 
         <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/floatingCityNameContainer"
+            android:id="@+id/floating_city_info_container"
             android:layout_width="0dp"
             android:layout_height="154dp"
             android:layout_margin="16dp"
@@ -40,6 +40,7 @@
             app:layout_constraintStart_toStartOf="parent">
 
             <com.google.android.material.button.MaterialButton
+                android:id="@+id/openCityWeatherFragmentButton"
                 style="@style/Widget.MaterialComponents.Button.OutlinedButton"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
