+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,5 @@
     <string name="getting_your_location">getting your current location wait..</string>
     <string name="something_went_wrong">something went wrong</string>
     <string name="show_weather">Show Weather</string>
+    <string name="googleServiceNotFound">can\'t find city name, google service is needed.</string>
 </resources>
